completed

+++ Develop/routes/api/product-routes.js
@@ -73,39 +73,50 @@ router.post('/', (req, res) => {
         });
 });
 
+// update product
 router.put('/:id', (req, res) => {
+    // update product data
     Product.update(req.body, {
-        where: {
-            id: req.params.id,
-        },
+      where: {
+        id: req.params.id,
+      },
     })
-        .then((product) => {
-            return ProductTag.findAll({ where: { product_id: req.params.id } });
-        })
-        .then((productTags) => {
+      .then((product) => {
+        if (req.body.tagIds && req.body.tagIds.length) {
+          
+          ProductTag.findAll({
+            where: { product_id: req.params.id }
+          }).then((productTags) => {
+            // create filtered list of new tag_ids
             const productTagIds = productTags.map(({ tag_id }) => tag_id);
             const newProductTags = req.body.tagIds
-                .filter((tag_id) => !productTagIds.includes(tag_id))
-                .map((tag_id) => {
-                    return {
-                        product_id: req.params.id,
-                        tag_id,
-                    };
-                });
+            .filter((tag_id) => !productTagIds.includes(tag_id))
+            .map((tag_id) => {
+              return {
+                product_id: req.params.id,
+                tag_id,
+              };
+            });
+  
+              // figure out which ones to remove
             const productTagsToRemove = productTags
-                .filter(({ tag_id }) => !req.body.tagIds.includes(tag_id))
-                .map(({ id }) => id);
-
+            .filter(({ tag_id }) => !req.body.tagIds.includes(tag_id))
+            .map(({ id }) => id);
+                    // run both actions
             return Promise.all([
-                ProductTag.destroy({ where: { id: productTagsToRemove } }),
-                ProductTag.bulkCreate(newProductTags),
+              ProductTag.destroy({ where: { id: productTagsToRemove } }),
+              ProductTag.bulkCreate(newProductTags),
             ]);
-        })
-        .then((updatedProductTags) => res.json(updatedProductTags))
-        .catch((err) => {
-            res.status(400).json(err);
-        });
-});
+          });
+        }
+  
+        return res.json(product);
+      })
+      .catch((err) => {
+        // console.log(err);
+        res.status(400).json(err);
+      });
+  });
 
 router.delete('/:id', (req, res) => {
     Product.destroy({
